[C-3378] Improve web-app hierarchy
