+++ packages/web/index.html
@@ -191,7 +191,7 @@
   <body>
     <noscript> You need to enable JavaScript to run this app. </noscript>
     <div id="root"></div>
-    <script type="module" src="/src/index.jsx"></script>
+    <script type="module" src="/src/index.tsx"></script>
     <div id="recaptcha-hidden" style="display: none"></div>
 
     <!-- start Meta Pixel -->
