+++ packages/web/src/app/AppErrorBoundary.tsx
