+++ packages/web/src/app/web-player/WebPlayer.test.tsx
@@ -1,5 +1,3 @@
-import { createRef, MutableRefObject } from 'react'
-
 import { ConnectedRouter } from 'connected-react-router'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
@@ -8,7 +6,7 @@ import { describe, it, vitest } from 'vitest'
 import { store } from 'store/configureStore'
 import history from 'utils/history'
 
-import App from './App'
+import WebPlayer from './WebPlayer'
 
 vitest.mock('jimp/es', () => null)
 vitest.mock('./visualizer/Visualizer', () => () => null)
@@ -31,14 +29,10 @@ vitest.mock('services/solana-client/SolanaClient', () => ({
 describe('smoke test', () => {
   it('renders without crashing', () => {
     const rootNode = document.createElement('div')
-    const mainContentRef =
-      createRef<HTMLDivElement | null>() as MutableRefObject<
-        HTMLDivElement | undefined
-      >
     ReactDOM.render(
       <Provider store={store}>
         <ConnectedRouter history={history}>
-          <App mainContentRef={mainContentRef} />
+          <WebPlayer />
         </ConnectedRouter>
       </Provider>,
       rootNode
