+++ packages/web/src/components/toast/ToastContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useCallback } from 'react'
+import { createContext, ReactNode, useCallback } from 'react'
 
 import { CommonState, toastActions } from '@audius/common'
 import { useDispatch, useSelector } from 'react-redux'
@@ -48,7 +48,7 @@ export const ToastContext = createContext<ToastContextProps>({
   toast: () => {}
 })
 
-export const ToastContextProvider = (props: { children: JSX.Element }) => {
+export const ToastContextProvider = (props: { children: ReactNode }) => {
   const toasts = useSelector((state: CommonState) => state.ui.toast.toasts)
   const dispatch = useDispatch()
