+++ packages/web/src/pages/sign-on/components/mobile/FollowPage.tsx
@@ -4,6 +4,7 @@ import { ID, User } from '@audius/common'
 import { Button, ButtonType, IconArrow } from '@audius/stems'
 import cn from 'classnames'
 
+import { MAIN_CONTENT_ID } from 'app/web-player/WebPlayer'
 import IconWand from 'assets/img/iconWand.svg'
 import {
   FollowArtistsCategory,
@@ -11,7 +12,6 @@ import {
 } from 'common/store/pages/signon/types'
 import UserCard from 'components/card/UserCard'
 import SelectablePills from 'components/selectable-pill/SelectablePills'
-import { MAIN_CONTENT_ID } from 'pages/App'
 
 import styles from './FollowPage.module.css'
