+++ packages/web/src/public-site/index.ts
@@ -1,0 +1,1 @@
+export { PublicSite as default } from './PublicSite'
