+++ packages/web/src/public-site/pages/download-page/DownloadPage.tsx
@@ -12,7 +12,7 @@ import IconDownloadMobile from 'assets/img/publicSite/downloadMobile.svg'
 import { CookieBanner } from 'components/cookie-banner/CookieBanner'
 import Footer from 'components/public-site/Footer'
 import NavBanner from 'components/public-site/NavBanner'
-import CTAStartListening from 'pages/landing-page/components/CTAStartListening'
+import CTAStartListening from 'public-site/pages/landing-page/components/CTAStartListening'
 import DownloadApp from 'services/download-app/DownloadApp'
 import { getIOSAppLink } from 'utils/appLinks'
 import { getOS } from 'utils/clientUtil'
