+++ packages/web/src/public-site/pages/landing-page/components/ArtistInvestors.tsx
