+++ packages/web/src/public-site/pages/landing-page/components/CaseStudies.tsx
