+++ packages/web/src/public-site/pages/landing-page/components/Description.tsx
