+++ packages/web/src/public-site/pages/landing-page/components/FeaturedContent.module.css
